requirements.txt
Django==1.7.3
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-debug-toolbar==1.2.2
django-enumfield==1.1.2
django-registration-redux==1.1
django-sslify==0.2.5
django-timezone-field==1.1
django-toolbelt==0.0.1
django-widget-tweaks==1.3
psycopg2==2.5.4
pytz==2014.10
sh==1.11
sqlparse==0.1.14
static3==0.5.1
uWSGI==2.0.9
whitenoise==1.0.6
wsgiref==0.1.2
